requirements.txt
matplotlib
numpy
Pillow
mediapipe
opencv-python
cvzone
torch
torchvision